+++ django_structlog/signals.py
@@ -25,6 +25,7 @@ bind_extra_request_finished_metadata = django.dispatch.Signal()
 
 :param logger: the logger
 :param response: the response resulting of the request
+:param log_kwargs: dictionary of log metadata for the ``request_finished`` event. It contains ``request`` and ``code`` keys. You may modify it to add extra information.
 
 >>> from django.contrib.sites.shortcuts import get_current_site
 >>> from django.dispatch import receiver
@@ -32,7 +33,7 @@ bind_extra_request_finished_metadata = django.dispatch.Signal()
 >>> import structlog
 >>>
 >>> @receiver(signals.bind_extra_request_finished_metadata)
-... def bind_domain(request, logger, response, **kwargs):
+... def bind_domain(request, logger, response, log_kwargs, **kwargs):
 ...     current_site = get_current_site(request)
 ...     structlog.contextvars.bind_contextvars(domain=current_site.domain)
 
@@ -43,6 +44,7 @@ bind_extra_request_failed_metadata = django.dispatch.Signal()
 
 :param logger: the logger
 :param exception: the exception resulting of the request
+:param log_kwargs: dictionary of log metadata for the ``request_failed`` event. It contains ``request`` and ``code`` keys. You may modify it to add extra information.
 
 >>> from django.contrib.sites.shortcuts import get_current_site
 >>> from django.dispatch import receiver
@@ -50,7 +52,7 @@ bind_extra_request_failed_metadata = django.dispatch.Signal()
 >>> import structlog
 >>>
 >>> @receiver(signals.bind_extra_request_failed_metadata)
-... def bind_domain(request, logger, exception, **kwargs):
+... def bind_domain(request, logger, exception, log_kwargs, **kwargs):
 ...     current_site = get_current_site(request)
 ...     structlog.contextvars.bind_contextvars(domain=current_site.domain)